Add unit tests for CLI and multiple image loading functionality
