+++ tests/unit/test_cli.py
